I2C reading and writing for CTA boards FE 0.  fix bug in I2C parameter multiple read.  function to retrieve the board type as configured in HWDescription.xml.  dump the parameter name when an exception is raised while reading or writing uhal parameters. address file for CTA boards

+++ Utils/Data.h
@@ -81,7 +81,7 @@ namespace Ph2_HwInterface
 			return ( ( fCurrentEvent >= fEventList.size() ) ? nullptr : fEventList.at( fCurrentEvent++ ) );
 		}
 		const Event* GetEvent( const BeBoard* pBoard, int i ) const {
-			return ( ( i >= fEventList.size() ) ? nullptr : fEventList.at( i ) );
+			return ( ( i >= (int)fEventList.size() ) ? nullptr : fEventList.at( i ) );
 		}
 		const std::vector<Event*>& GetEvents( const BeBoard* pBoard ) const {
 			return fEventList;
